Create README.md
I need that code to perform gases measurements (air, fumes) inside of duct (shape: rectangle or circle) with use of a Prandtl tube. I am using the Log-Chebyshev method during the measurements. I enter the diameter of the pipeline into the program and specify the number of measurement points I want to make. The program asks me for an additional data: gas temperature, static pressure in the duct, ambient pressure and the Prandtl tube coefficient. Then the program showes me the coordinates of the measuring points, and I enter the pressure measurement result for a given point. Program use all of entered data to convert measured pressures into velocities and calculate average velocity of gas inside of the duct. The last step of calculation is to convert the average velocity of gas inside the duct into the gas flow. User can choose unit of printed result.